# server/docs/swagger.yaml
basePath: /
definitions:
...
  dtos.BadgeDTO:
    properties:
      id:
        type: string
      name:
        minLength: 3
        type: string
      nbRequirementEvent:
        minimum: 1
        type: integer
      svg:
        type: string
      type:
        enum:
        - EVENT
        - USER
        - MESSAGE
        type: string
    required:
    - name
    - nbRequirementEvent
    - svg
    - type
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the EduMeet project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: EduMeet API Documentation
  version: "1.0"
paths:
...
  /badge:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all badges
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: List of badges'
          schema:
            items:
              $ref: '#/definitions/dtos.BadgeDTO'
            type: array
        "404":
          description: 'Error: Badges not found'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all badges
      tags:
      - Badge
    post:
      consumes:
      - application/json
      description: Create a new badge and add it to the system
      parameters:
      - description: Badge Data
        in: body
        name: badge
        required: true
        schema:
          $ref: '#/definitions/dtos.BadgeDTO'
      produces:
      - application/json
      responses:
        "201":
          description: 'Success: Badge created'
          schema:
            $ref: '#/definitions/dtos.BadgeDTO'
        "400":
          description: 'Error: Invalid data'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'Error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: 'Error: Validation failed'
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new badge
      tags:
      - Badge
  /badge/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a badge from the system by its ID
      parameters:
      - description: Badge ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 'Success: No Content'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Error: Invalid ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'Error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'Error: Forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'Error: Badge not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a badge by ID
      tags:
      - Badge
    get:
      consumes:
      - application/json
      description: Retrieve a badge by its ID
      parameters:
      - description: Badge ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Badge data'
          schema:
            $ref: '#/definitions/dtos.BadgeDTO'
        "400":
          description: 'Error: Invalid ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'Error: Badge not found'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a single badge by ID
      tags:
      - Badge
    put:
      consumes:
      - application/json
      description: Update the details of an existing badge
      parameters:
      - description: Badge ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Badge Data
        in: body
        name: badge
        required: true
        schema:
          $ref: '#/definitions/dtos.BadgeDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Badge updated'
          schema:
            $ref: '#/definitions/dtos.BadgeDTO'
        "400":
          description: 'Error: Invalid ID or data'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'Error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'Error: Badge not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a badge by ID
      tags:
      - Badge
...
  /document:
    post:
      consumes:
      - multipart/form-data
      description: Create a new document
      parameters:
      - description: File
        in: formData
        name: file
        required: true
        type: file
      - description: Event ID
        in: formData
        name: event_id
        required: true
        type: string
      - description: Message ID
        in: formData
        name: message_id
        required: true
        type: string
      - description: Document Type
        in: formData
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: 'Success: Document created'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Missing or invalid data'
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: 'Unprocessable Entity: Validation errors'
          schema:
            additionalProperties: true
            type: object
      summary: Create Document
      tags:
      - Document
  /document/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a document by its ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 'Success: Document deleted'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'Not Found: Document not found'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete Document
      tags:
      - Document
    get:
      consumes:
      - application/json
      description: Retrieve a document by its ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Document file'
          schema:
            type: file
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'Not Found: Document not found'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Document
      tags:
      - Document
  /document/details/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve details of a document by its ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Document details'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'Not Found: Document not found'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Document Details
      tags:
      - Document
  /event/{id}/documents:
    get:
      consumes:
      - application/json
      description: Retrieve all documents related to an event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Event documents'
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'Not Found: Documents not found'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Event Documents
      tags:
      - Document
...
swagger: "2.0"